+++ applications/MachineStateManager/src/socket_interface.cpp
@@ -25,7 +25,7 @@ int SocketInterface::listen(int sockfd, int backlog)
   return ::listen(sockfd, backlog);
 }
 
-int SocketInterface::recv(int sockfd, void *buf, size_t len, int flags)
+ssize_t SocketInterface::recv(int sockfd, void *buf, size_t len, int flags)
 {
   return ::recv(sockfd, buf, len, flags);
 }
