+++ common/include/i_socket_interface.hpp
@@ -11,7 +11,7 @@ public:
   virtual int socket(int domain, int type, int protocol) = 0;
   virtual int accept(int sockfd, struct sockaddr* addr, socklen_t* addrlen) = 0;
   virtual int listen(int sockfd, int backlog) = 0;
-  virtual int recv(int sockfd, void *buf, size_t len, int flags) = 0;
+  virtual ssize_t recv(int sockfd, void *buf, size_t len, int flags) = 0;
 
   virtual ~ISocketInterface() = default;
 };
